+++ Common/src/main/java/ca/lukegrahamlandry/lib/data/nbt/NbtDataWrapper.java
@@ -9,6 +9,7 @@
 
 package ca.lukegrahamlandry.lib.data.nbt;
 
+import ca.lukegrahamlandry.lib.base.WrappedData;
 import ca.lukegrahamlandry.lib.base.json.JsonHelper;
 import com.google.gson.Gson;
 import com.google.gson.JsonSyntaxException;
@@ -22,44 +23,39 @@ import java.lang.reflect.InvocationTargetException;
 import java.util.HashMap;
 import java.util.Map;
 
-public abstract class NbtDataWrapper<O, V> {
+public abstract class NbtDataWrapper<O, V, S extends NbtDataWrapper<O, V, S>> extends WrappedData<V, S> {
     public static <V> ItemStackDataWrapper<V> itemStack(Class<V> clazz){
         return new ItemStackDataWrapper<>(TypeToken.get(clazz));
     }
 
-    public <W extends NbtDataWrapper<O, V>> W named(String name){
+    public S named(String name){
         this.name = name;
-        this.createLogger();
-        return (W) this;
+        this.updateLogger();
+        return (S) this;
     }
 
-    public <W extends NbtDataWrapper<O, V>> W named(ResourceLocation name){
+    public S named(ResourceLocation name){
         return this.named(name.toString());
-    }
-
-    public <W extends NbtDataWrapper<O, V>> W withGson(Gson gson){
-        this.gson = gson;
-        return (W) this;
     }
 
     // API
 
     public V get(O obj){
-        int hashcode = getHashCode(obj);
+        int hashcode = this.getHashCode(obj);
         if (previousDataObjects.containsKey(hashcode)) return previousDataObjects.get(hashcode);
 
         V value;
-        if (getSharedTag(obj).contains(getFullTagKey())) {
-            String data = getSharedTag(obj).getString(getFullTagKey());
+        if (this.getSharedTag(obj).contains(this.getFullTagKey())) {
+            String data = this.getSharedTag(obj).getString(this.getFullTagKey());
             try {
-                value = getGson().fromJson(data, valueType.getType());
+                value = this.getGson().fromJson(data, this.getValueType());
             } catch (JsonSyntaxException e){
-                this.logger.error("Using default. Failed to parse data: " + data);
+                this.getLogger().error("Using default. Failed to parse data: " + data);
                 e.printStackTrace();
-                value = getDefaultInstance();
+                value = this.getDefaultValue();
             }
         } else {
-            value = getDefaultInstance();
+            value = this.getDefaultValue();
         }
 
         previousDataObjects.put(getHashCode(obj), value);
@@ -67,56 +63,39 @@ public abstract class NbtDataWrapper<O, V> {
     }
 
     public void set(O obj, V value){
-        String data = getGson().toJson(value);
-        getSharedTag(obj).putString(getFullTagKey(), data);
-        previousDataObjects.put(getHashCode(obj), value);
+        String data = this.getGson().toJson(value);
+        this.getSharedTag(obj).putString(this.getFullTagKey(), data);
+        previousDataObjects.put(this.getHashCode(obj), value);
     }
 
     public void setDirty(O obj){
-        set(obj, previousDataObjects.get(getHashCode(obj)));
+        this.set(obj, previousDataObjects.get(this.getHashCode(obj)));
     }
 
     public void remove(O obj){
-        previousDataObjects.remove(getHashCode(obj));
-        getSharedTag(obj).remove(getFullTagKey());
+        previousDataObjects.remove(this.getHashCode(obj));
+        this.getSharedTag(obj).remove(this.getFullTagKey());
     }
 
     // IMPL
 
     public static final String PARENT_TAG_KEY = "WrapperLib";
     Map<Integer, V> previousDataObjects = new HashMap<>();
-    protected TypeToken<V> valueType;
-    protected Logger logger; // TODO
     private String name;
-    private Gson gson;
-    public NbtDataWrapper(TypeToken<V> valueType){
-        this.valueType = valueType;
-        this.named(valueType.getRawType().getSimpleName());
+    public NbtDataWrapper(TypeToken<V> type){
+        super(type);
+        this.named(type.getRawType().getSimpleName());
     }
 
     String getFullTagKey(){
         return this.name;
     }
 
-    private V getDefaultInstance(){
-        try {
-            return (V) this.valueType.getRawType().getConstructor().newInstance();
-        } catch (NoSuchMethodException | InvocationTargetException | InstantiationException | IllegalAccessException e) {
-            this.logger.error(this.valueType.getRawType().getName() + " does not have a public parameterless constructor");
-            throw new RuntimeException(this.valueType.getRawType().getName() + " does not have a public parameterless constructor", e);
-        }
-    }
-
     protected abstract CompoundTag getSharedTag(O obj);
 
     protected abstract int getHashCode(O obj);
 
-    public Gson getGson(){
-        return this.gson == null ? JsonHelper.get() : this.gson;
-    }
-
-    protected void createLogger(){
-        String id = this.getClass().getName() + ": " + this.name;
-        this.logger = LoggerFactory.getLogger(id);
+    protected String getAdditionalLoggerId(){
+        return this.name;
     }
 }