+++ Common/src/main/java/ca/lukegrahamlandry/lib/resources/ResourcesWrapper.java
@@ -11,28 +11,24 @@ package ca.lukegrahamlandry.lib.resources;
 
 import ca.lukegrahamlandry.lib.base.Available;
 import ca.lukegrahamlandry.lib.base.InternalUseOnly;
-import ca.lukegrahamlandry.lib.base.json.JsonHelper;
-import ca.lukegrahamlandry.lib.config.ConfigWrapper;
+import ca.lukegrahamlandry.lib.base.WrappedData;
 import ca.lukegrahamlandry.lib.helper.PlatformHelper;
-import com.google.gson.Gson;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonSyntaxException;
 import com.google.gson.reflect.TypeToken;
 import dev.architectury.injectables.annotations.ExpectPlatform;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.server.MinecraftServer;
+import net.minecraft.server.packs.resources.PreparableReloadListener;
 import net.minecraft.server.packs.resources.Resource;
 import net.minecraft.server.packs.resources.ResourceManager;
-import net.minecraft.server.packs.resources.SimplePreparableReloadListener;
 import net.minecraft.util.profiling.ProfilerFiller;
-import org.slf4j.Logger;
 
 import java.io.BufferedReader;
 import java.util.*;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.Executor;
 
-public class ResourcesWrapper<T> extends SimplePreparableReloadListener<Map<ResourceLocation, List<JsonElement>>> {
-    static MinecraftServer server = null;
-
+public class ResourcesWrapper<T> extends WrappedData<T, ResourcesWrapper<T>> implements PreparableReloadListener {
     /**
      * Load information from a data pack on the logical server.
      */
@@ -71,19 +67,21 @@ public class ResourcesWrapper<T> extends SimplePreparableReloadListener<Map<Reso
         return this;
     }
 
-    public ResourcesWrapper<T> withGson(Gson gson){
-        this.gson = gson;
-        return this;
-    }
-
     // API
 
     public Set<Map.Entry<ResourceLocation, T>> entrySet(){
+        if (this.data == null){
+            this.getLogger().error("Cannot call ResourcesWrapper#entrySet before resource listeners are loaded");
+            return null;
+        }
         return data.entrySet();
     }
 
     public T get(ResourceLocation id){
-        if (data == null) return null;
+        if (this.data == null){
+            this.getLogger().error("Cannot call ResourcesWrapper#get before resource listeners are loaded");
+            return null;
+        }
         return data.get(id);
     }
 
@@ -94,19 +92,16 @@ public class ResourcesWrapper<T> extends SimplePreparableReloadListener<Map<Reso
     // IMPL
 
     static List<ResourcesWrapper<?>> ALL = new ArrayList<>();
-    TypeToken<T> valueType;
     public final String directory;
     public final boolean isServerSide;
-    Map<ResourceLocation, T> data;
-    Logger logger;
+    Map<ResourceLocation, T> data = null;
     protected String suffix = ".json";
-    private Gson gson = JsonHelper.get();
     boolean shouldSync = false;
     protected Runnable onLoadAction = () -> {};
     protected Runnable onReceiveSyncAction = () -> {};
     protected MergeRule<T> mergeRule = (resources) -> resources.get(resources.size() - 1);  // TODO: make sure I guessed the priority order right
     public ResourcesWrapper(TypeToken<T> valueType, String directory, boolean isServerSide){
-        this.valueType = valueType;
+        super(valueType);
         this.directory = directory;
         this.isServerSide = isServerSide;
         if (!this.isServerSide && Available.PLATFORM_HELPER.get() && PlatformHelper.isDedicatedServer()) return;
@@ -114,17 +109,13 @@ public class ResourcesWrapper<T> extends SimplePreparableReloadListener<Map<Reso
         registerResourceListener(this);
     }
 
-    public Gson getGson(){
-        return this.gson;
-    }
 
     /**
      * For each filename in our directory of each pack we retrieve the list of files.
      * Each file comes from a different pack. They all get added to a list under their resource location.
      * This map with all the entries from every pack gets returned and later passed to PackWrapper#apply.
      */
-    @Override
-    protected Map<ResourceLocation, List<JsonElement>> prepare(ResourceManager resourceManager, ProfilerFiller profilerFiller) {
+    protected Map<ResourceLocation, List<JsonElement>> prepare(ResourceManager resourceManager, ProfilerFiller profiler) {
         Map<ResourceLocation, List<JsonElement>> results = new HashMap<>();
         for (Map.Entry<ResourceLocation, List<Resource>> entry : resourceManager.listResourceStacks(this.directory, file -> file.getPath().endsWith(suffix)).entrySet()) {
             String actualPath = entry.getKey().getPath().substring(this.directory.length() + 1, entry.getKey().getPath().length() - suffix.length());
@@ -134,10 +125,10 @@ public class ResourcesWrapper<T> extends SimplePreparableReloadListener<Map<Reso
                 try (BufferedReader reader = resource.openAsReader()) {
                     values.add(this.getGson().fromJson(reader, JsonElement.class));
                 } catch (JsonSyntaxException e) {
-                    this.logger.error("Failed to parse json for " + id + " from pack " + resource.sourcePackId());
+                    this.getLogger().error("Failed to parse json for " + id + " from pack " + resource.sourcePackId());
                     e.printStackTrace();
                 } catch (Exception e) {
-                    this.logger.error("Failed to read " + id + " from pack " + resource.sourcePackId());
+                    this.getLogger().error("Failed to read " + id + " from pack " + resource.sourcePackId());
                     e.printStackTrace();
                 }
             }
@@ -150,8 +141,7 @@ public class ResourcesWrapper<T> extends SimplePreparableReloadListener<Map<Reso
      * For each resource location, all the json elements get parsed into our data type.
      * The resultant list is passed to the merge rule and its result is saved.
      */
-    @Override
-    protected void apply(Map<ResourceLocation, List<JsonElement>> resources, ResourceManager resourceManager, ProfilerFiller profilerFiller) {
+    protected void apply(Map<ResourceLocation, List<JsonElement>> resources, ResourceManager resourceManager, ProfilerFiller profiler) {
         this.data = new HashMap<>();
         for (Map.Entry<ResourceLocation, List<JsonElement>> entry : resources.entrySet()){
             ResourceLocation id = entry.getKey();
@@ -159,14 +149,23 @@ public class ResourcesWrapper<T> extends SimplePreparableReloadListener<Map<Reso
             List<T> values = new ArrayList<>();
 
             for (JsonElement json : resourceStack){
-                values.add(this.getGson().fromJson(json, this.valueType.getType()));
+                values.add(this.getGson().fromJson(json, this.getValueType()));
             }
 
             T finalValue = this.mergeRule.merge(values);
             this.data.put(id, finalValue);
         }
         this.onLoadAction.run();
-        if (this.shouldSync && server != null) new DataPackSyncMessage(this).sendToAllClients();
+        if (this.shouldSync) new DataPackSyncMessage(this).sendToAllClients();
+        this.getLogger().info("Loaded " + this.data.size() + " entries.");
+    }
+
+    // SimplePreparableReloadListener
+    @Override
+    public CompletableFuture<Void> reload(PreparableReloadListener.PreparationBarrier preparationBarrier, ResourceManager resources, ProfilerFiller prepareProfiler, ProfilerFiller applyProfiler, Executor prepareExecutor, Executor applyExecutor) {
+        return CompletableFuture.supplyAsync(() -> this.prepare(resources, prepareProfiler), prepareExecutor)
+                        .thenCompose(preparationBarrier::wait)
+                        .thenAcceptAsync(object -> this.apply(object, resources, applyProfiler), applyExecutor);
     }
 
     @InternalUseOnly